Create a simply html bar chart

arent div is set to 100% width and height

Usage

create_html_vertical_bar(
  bar_height = 0.5,
  bar_color = "#4a6fa5",
  parent_div_height = "50px",
  parent_div_width = "100%"
)

Arguments

bar_height

a number between 0 and 1 that sets the height of the bar as a percentage of the total height of the container (the parent) div

parent_div_height

valid css unit for height of the parent div

parent_div_width

valid css unit for width of the parent div

Examples

create_html_vertical_bar(0.5)
#> <div style="width:100%;height:50px;margin:0;padding:0;">
#>   <div style="height:50%;background:transparent;margin:0;padding:0;"></div>
#>   <div style="height:50%;background:#4a6fa5;margin:0;padding:0;"></div>
#> </div>