function that takes a dataframe, with sequences in one column, and creates a gt table

splits the seq column into 1 per element

applies a color palette to amino acids

Usage

gtseq(
  df,
  seq_column,
  names_prefix = "pos_",
  rowname_col = NULL,
  groupname_col = dplyr::group_vars(df),
  process_md = FALSE,
  caption = NULL,
  rownames_to_stub = FALSE,
  row_group_as_column = FALSE,
  auto_align = TRUE,
  id = NULL,
  locale = getOption("gt.locale"),
  row_group.sep = getOption("gt.row_group.sep", " - ")
)

Arguments

df

a data frame

seq_column

the column that contains biological sequences

names_prefix

the prefix to add to each position column

rowname_col

Column(s) for row names/labels from data

vector<character> // default: NULL (optional)

The column name(s) in the input data table to use as row labels to be placed in the table stub. If multiple column names are provided, they will create a hierarchical stub layout where values from the leftmost column form the highest level of the hierarchy, proceeding to individual row identifiers in the rightmost column. If the rownames_to_stub option is TRUE then any column name provided to rowname_col will be ignored.

groupname_col

Column for group names/labels from data

scalar<character> // default: NULL (optional)

The column name in the input data table to use as group labels for generation of row groups. If the input data table has the grouped_df class (through use of dplyr::group_by() or associated group_by*() functions) then any input here is ignored.

process_md

Process Markdown in rowname_col and groupname_col

scalar<logical> // default: FALSE

Should the contents of the rowname_col and groupname_col be interpreted as Markdown? By default this won't happen.

caption

Table caption text

scalar<character> // default: NULL (optional)

An optional table caption to use for cross-referencing in R Markdown, Quarto, or bookdown.

rownames_to_stub

Use data frame row labels in the stub

scalar<logical> // default: FALSE

An option to take rownames from the input data table (should they be available) as row labels in the display table stub.

row_group_as_column

Mode for displaying row group labels in the stub

scalar<logical> // default: FALSE

An option that alters the display of row group labels. By default this is FALSE and row group labels will appear in dedicated rows above their respective groups of rows. If TRUE row group labels will occupy a secondary column in the table stub.

auto_align

Automatic alignment of column values and labels

scalar<logical> // default: TRUE

Optionally have column data be aligned depending on the content contained in each column of the input data. Internally, this calls cols_align(align = "auto") for all columns.

id

The table ID

scalar<character> // default: NULL (optional)

By default (with NULL) this will be a random, ten-letter ID as generated by using random_id(). A custom table ID can be used here by providing a character value.

locale

Locale identifier

scalar<character> // default: getOption("gt.locale") (optional)

An optional locale identifier that can be set as the default locale for all functions that take a locale argument. Examples include "en" for English (United States) and "fr" for French (France). We can call info_locales() as a useful reference for all of the locales that are supported. If set, options(gt.locale) is also consulted.

row_group.sep

Separator text for multiple row group labels

scalar<character> // default: getOption("gt.row_group.sep", " - ")

The separator to use between consecutive group names (a possibility when providing data as a grouped_df with multiple groups) in the displayed row group label.

Value

a gt table object

Examples

msa |> gtseq(seq_column = seq)

  

      	group
      	name
      	start
      	
      	
      	
      	
      	5
      	
      	
      	
      	
      	10
      	
      	
      	
      	
      	15
      	
      	
      	
      	
      	20
      	
      	
      	
      	
      	25
      	
      	
      	
      	
      	30
      	
      	
      	
      	
      	35
      	
      	
      	
      	
      	40
      	
      	
      	
      	
      	45
      	
      	
      	
      	
      	50
      	
      	
      	
      	
      	55
      	
      	
      	
      	
      	60
      	
      	
      	
      	
      	65
      	
      	
      	
      	
      	70
      	
      	
      	
      	
      	75
      	
      	
      	
      	
      	80
      	
      	
      	
      	
      	85
      	
      	
      	
      	
      	90
      	
      	
      	
      	
      	95
    

	beta
	OC43
	1225
	T
	S
	I
	P
	N
	L
	P
	D
	F
	K
	E
	E
	L
	D
	Q
	W
	F
	K
	N
	Q
	T
	S
	-
	V
	A
	P
	D
	L
	S
	L
	D
	Y
	-
	-
	I
	N
	V
	T
	F
	L
	D
	L
	Q
	V
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	E
	M
	N
	R
	L
	Q
	E
	A
	I
	K
	V
	L
	N
	Q
	S
	Y
	I
	N
	L
	K
	D
	I
	G
	T
	Y
	E
	Y
	Y
	V
	K
	W
	P
	W
	Y
	V
	W
	L

	beta
	MHV
	1191
	T
	S
	I
	P
	N
	P
	P
	D
	F
	K
	E
	E
	L
	D
	Q
	W
	F
	K
	K
	Q
	T
	S
	-
	I
	A
	P
	D
	L
	S
	L
	D
	F
	E
	K
	L
	N
	V
	T
	L
	L
	D
	L
	T
	Y
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	E
	M
	N
	R
	I
	Q
	D
	A
	I
	K
	K
	L
	N
	E
	S
	Y
	I
	N
	L
	K
	E
	V
	G
	T
	Y
	E
	M
	Y
	V
	K
	W
	P
	W
	Y
	V
	W
	L

	beta
	HKU1
	1226
	H
	S
	V
	P
	K
	L
	S
	D
	F
	E
	S
	E
	L
	S
	H
	W
	F
	K
	N
	Q
	T
	S
	-
	I
	A
	P
	N
	L
	T
	L
	N
	L
	H
	T
	I
	N
	A
	T
	F
	L
	D
	L
	Y
	Y
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	E
	M
	N
	L
	I
	Q
	E
	S
	K
	L
	S
	L
	N
	N
	S
	Y
	I
	N
	L
	K
	D
	I
	G
	T
	Y
	E
	M
	Y
	V
	K
	W
	P
	W
	Y
	V
	W
	L

	beta
	SARS
	1122
	P
	L
	Q
	P
	E
	L
	D
	S
	F
	K
	E
	E
	L
	D
	K
	Y
	F
	K
	N
	H
	T
	S
	-
	P
	D
	V
	D
	F
	G
	-
	D
	I
	S
	G
	I
	N
	A
	S
	V
	V
	N
	I
	Q
	K
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	E
	I
	D
	R
	L
	N
	E
	V
	A
	K
	N
	L
	N
	E
	S
	L
	I
	D
	L
	Q
	E
	L
	G
	K
	Y
	E
	Q
	Y
	I
	K
	W
	P
	W
	Y
	V
	W
	L

	beta
	SARS2
	1140
	P
	L
	Q
	P
	E
	L
	D
	S
	F
	K
	E
	E
	L
	D
	K
	Y
	F
	K
	N
	H
	T
	S
	-
	P
	D
	V
	D
	L
	G
	-
	D
	I
	S
	G
	I
	N
	A
	S
	V
	V
	N
	I
	Q
	K
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	E
	I
	D
	R
	L
	N
	E
	V
	A
	K
	N
	L
	N
	E
	S
	Y
	I
	D
	L
	K
	E
	L
	G
	N
	Y
	T
	Y
	Y
	N
	K
	W
	P
	W
	Y
	I
	W
	L

	beta
	MERS
	1223
	L
	G
	N
	S
	T
	G
	I
	D
	F
	Q
	D
	E
	L
	D
	E
	Y
	F
	K
	N
	V
	S
	T
	-
	S
	I
	P
	N
	F
	G
	-
	S
	L
	T
	Q
	I
	N
	T
	T
	L
	L
	D
	L
	T
	Y
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	-
	E
	M
	L
	S
	L
	Q
	Q
	V
	V
	K
	A
	L
	N
	E
	S
	Y
	I
	D
	L
	K
	E
	L
	G
	N
	Y
	T
	Y
	Y
	N
	K
	W
	P
	W
	Y
	I
	W
	L

	alpha
	229E
	1025
	T
	I
	V
	P
	E
	Y
	I
	D
	V
	N
	K
	T
	L
	Q
	E
	L
	S
	Y
	K
	L
	P
	N
	Y
	T
	V
	P
	D
	L
	-
	-
	V
	V
	E
	Q
	Y
	N
	Q
	T
	I
	L
	N
	L
	T
	S
	E
	I
	S
	T
	L
	E
	N
	K
	S
	A
	E
	L
	N
	Y
	T
	V
	Q
	K
	L
	Q
	T
	L
	I
	D
	N
	I
	N
	S
	T
	L
	V
	D
	L
	K
	W
	L
	N
	R
	V
	E
	T
	Y
	I
	K
	W
	P
	W
	Y
	V
	W
	V

	alpha
	NL63
	1208
	T
	V
	I
	P
	D
	Y
	V
	D
	V
	N
	K
	T
	L
	Q
	E
	F
	A
	Q
	N
	L
	P
	K
	Y
	V
	K
	P
	N
	F
	-
	-
	D
	L
	T
	P
	F
	N
	L
	T
	Y
	L
	N
	L
	S
	S
	E
	L
	K
	Q
	L
	E
	A
	K
	T
	A
	S
	L
	F
	Q
	T
	T
	V
	E
	L
	Q
	G
	L
	I
	D
	Q
	I
	N
	S
	T
	Y
	V
	D
	L
	K
	L
	L
	N
	R
	F
	E
	N
	Y
	I
	K
	W
	P
	W
	Y
	V
	W
	V

Consensus

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

  

    

    

  

Sequence

	T
	S
	I
	P
	E
	L
	D
	D
	F
	K
	E
	E
	L
	D
	E
	W
	F
	K
	N
	Q
	T
	S
	–
	I
	A
	P
	D
	L
	G
	–
	D
	L
	E
	G
	I
	N
	A
	T
	F
	L
	D
	L
	Q
	Y
	–
	–
	–
	–
	–
	–
	–
	–
	–
	–
	–
	–
	–
	–
	E
	M
	N
	R
	L
	Q
	E
	V
	I
	K
	N
	L
	N
	E
	S
	Y
	I
	D
	L
	K
	E
	L
	G
	T
	Y
	E
	Y
	Y
	I
	K
	W
	P
	W
	Y
	V
	W
	L